added user & group routes

+++ src/main/java/dat/dtos/GroupDTO.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -14,7 +13,7 @@ import java.util.stream.Collectors;
 @Getter
 @NoArgsConstructor
 public class GroupDTO {
-    private Long id;
+    private int id;
     private String name;
     private String description;
     private Set<Integer> memberIds = new HashSet<>();
